requirements.txt
Flask
datetime
uuid
PyJWT
kafka-python
pymongo
flask_pymongo